[Warlock] Demonology AoE APL Improvements (#10185)
* [Warlock][Demonology] Improves AOE Rotation around TWW2 [4]PC Bonus

https://www.raidbots.com/simbot/report/iaGzMwDH3XmRVmxSqXXF3A 

Slight upgrade in damage by taking advantage of the TWWS2 4p bonus and increasing the amount of Hand of Guldan casts.

* [Warlock][Demonology] Update towards TWW2 Tierset/Implosion

* Update warlock.cpp

https://www.raidbots.com/simbot/report/6pFKMk3n5Ty2wzTcBEtWA5/simc

https://www.raidbots.com/simbot/report/cMZQG868soiVkaEYvSjgY8

One final change, conciliates the lower desire to implode while having tww2 tierset with a logic aimed towards  the  objective of demonic core generation   and keeps a separate line for the sake of implosions without the tierset. 

Still keeps target counts for implosions above 2.

* Update warlock.cpp