[Shaman] Introduce randomized targeting for chained spells
Chain Lightning and Lava Beam by default now have a 15% chance per jump
to shuffle the target into a new position in the target list, causing
new targets to appear in the jump list instead. A target can be shuffled
only once, and the primary target of the actor will never be shuffled,
nor will a new target become the "first" target of the jump sequence.

Target shuffling will only take place when there are more targets
available than there are jumps for the chained spell.

The option `shaman.chain_lightning_target_rng` [0..1] controls the proc
chance for the shuffling to occur.